Update dependencies and enhance WebSocket setup
This commit includes the update of several packages including 'aws-cdk-lib' to the version 2.147.1. Additionally, it establishes a new setup for WebSockets, including the integration of the 'dotenv' package for environment variable management. Changes in the DynamoDB table definition, Lambda function definition, and API Gateway configuration were introduced as part of these improvements.