feat: Add SQS integration for sending messages to connections
The code changes include adding the  dependency to the  file and implementing a new method  in the  class. Additionally, a new file  is added to handle SQS events and send messages to connections using the WebSocket API. This commit also updates the  class to include the definition of an SQS queue and grant necessary permissions to the Lambda functions.